Add beta support for Redhat distro family
* Tested only on centos

+++ install.sh
@@ -1,9 +1,31 @@
 #!/bin/bash
-sudo apt update && sudo apt -y upgrade && sudo apt install gcc
+set -e
+
+if [ -f /etc/redhat-release ]; then
+  export DISTRO="REDHAT"
+elif [ -f /etc/lsb-release ]; then
+  export DISTRO="DEBIAN"
+else
+  echo "Unknown distro - Exit"
+  exit 1
+fi
+
+if [ "$DISTRO" == "DEBIAN" ]; then
+  sudo apt install -y gcc git \
+  libncurses5-dev libgnome2-dev libgnomeui-dev \
+  libgtk2.0-dev libatk1.0-dev libbonoboui2-dev \
+  libcairo2-dev libx11-dev libxpm-dev libxt-dev python-dev \
+  python3-dev ruby-dev lua5.1 liblua5.1-dev libperl-dev git
+elif [ "$DISTRO" == "REDHAT" ]; then
+  sudo yum groupinstall -y 'Development tools'
+  sudo yum install -y gcc-c++ ncurses lua-devel ncurses-devel python-devel git python3 perl-devel perl-ExtUtils-Embed cmake python3-devel npm
+fi
+
 FILEDIR=$(cd "$(dirname "$0")"; pwd)
+
 rm -f ~/.bashrc
 rm -f ~/.vimrc
-rm -f ~/.vim
+rm -rf ~/.vim
 
 # Distribute bash/vim rcs
 cp "$FILEDIR/.bashrc" ~/.bashrc
@@ -15,17 +37,21 @@ if [ ! -d "~/.vim/bundle/Vundle.vim" ]; then
   git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 fi
 
-# Get files to compile vim
-sudo apt -y install libncurses5-dev libgnome2-dev libgnomeui-dev \
-  libgtk2.0-dev libatk1.0-dev libbonoboui2-dev \
-  libcairo2-dev libx11-dev libxpm-dev libxt-dev python-dev \
-  python3-dev ruby-dev lua5.1 liblua5.1-dev libperl-dev git
-
 # Remove vim
-sudo apt -y remove vim vim-runtime gvim
+if [ "$DISTRO" == "DEBIAN" ]; then
+  sudo apt -y remove vim vim-runtime gvim
+elif [ "$DISTRO" == "REDHAT" ]; then
+  sudo yum -y erase vim
+  rm -f /usr/bin/vim
+fi
 
-# Get python config dir name.
-LASTDIR="$(ls /usr/lib/python3.* | grep -E 'config-[0-9a-z]')"
+# Get python config dir name
+if [ "$DISTRO" == "REDHAT" ]; then
+  PYDIR="/usr/lib64/python3.6/config"
+elif [ "$DISTRO" == "DEBIAN" ]; then
+  PYDIR="/usr/lib/python3.5/config"
+fi
+
 git clone https://github.com/vim/vim.git
 cd vim
 
@@ -34,7 +60,7 @@ cd vim
             --enable-multibyte \
 	    --enable-rubyinterp=yes \
 	    --enable-python3interp=yes \
-	    --with-python3-config-dir=/usr/lib/python3.5/$LASTDIR \
+	    --with-python3-config-dir=$PYDIR \
 	    --enable-perlinterp=yes \
 	    --enable-luainterp=yes \
             --enable-gui=gtk2 \
@@ -44,32 +70,38 @@ make VIMRUNTIMEDIR=/usr/local/share/vim/vim82
 sudo make install
 
 # Let everyone know vim is installed
-sudo update-alternatives --install /usr/bin/editor editor /usr/local/bin/vim 1
-sudo update-alternatives --set editor /usr/local/bin/vim
-sudo update-alternatives --install /usr/bin/vi vi /usr/local/bin/vim 1
-sudo update-alternatives --set vi /usr/local/bin/vim
+if [ "$DISTRO" == "DEBIAN" ]; then
+  sudo update-alternatives --install /usr/bin/editor editor /usr/local/bin/vim 1
+  sudo update-alternatives --set editor /usr/local/bin/vim
+  sudo update-alternatives --install /usr/bin/vi vi /usr/local/bin/vim 1
+  sudo update-alternatives --set vi /usr/local/bin/vim
+fi
 
 # Get pathogen
 mkdir -p ~/.vim/autoload ~/.vim/bundle && \
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 
-# Get packages for color coded compilation
-sudo apt-get -y install build-essential libclang-3.9-dev libncurses-dev libz-dev cmake xz-utils libpthread-workqueue-dev
+if [ "$DISTRO" == "DEBIAN" ]; then
+  # Get packages for color coded compilation
+  sudo apt-get -y install build-essential libclang-3.9-dev libncurses-dev libz-dev cmake xz-utils libpthread-workqueue-dev
 
-# Get lua version
-LUAVERSION="$(vim --version | grep -o "lua[0-9].[0-9]")"
-sudo apt -y install lib$LUAVERSION-dev $LUAVERSION g++
+  # Get lua version
+  LUAVERSION="$(vim --version | grep -o "lua[0-9].[0-9]")"
+  sudo apt -y install lib$LUAVERSION-dev $LUAVERSION g++
+fi
 
 # Install plugins
 vim +PluginInstall +qall
 
-# Compile color coded
-cd ~/.vim/bundle/color_coded
-rm -f CMakeCache.txt
-mkdir build && cd build
-cmake ..
-make && make install
-make clean && make clean_clang
+if [ "$DISTRO" == "DEBIAN" ]; then
+  # Compile color coded
+  cd ~/.vim/bundle/color_coded
+  rm -f CMakeCache.txt
+  mkdir build && cd build
+  cmake ..
+  make && make install
+  make clean && make clean_clang
+fi
 
 # Install YCM & ask user whether C support or other lang
 cd ~/.vim/bundle/YouCompleteMe
